dynamically renders poll cards
