+++ src/features/polling/Poll.tsx
@@ -1,4 +1,6 @@
 import {Button} from "@mui/material";
+import Header from "../misc/Header";
+import PollOption from "./PollOption";
 
 const Poll = () => {
 
@@ -10,26 +12,13 @@ const Poll = () => {
         width: '80%'
     }
 
-    const cardStyle = {
-        backgroundColor: "white",
-        width: '25%',
-        padding: '.5%',
-        margin: '.5%',
-        display: 'inline-block'
-    }
-
     return (
         <div style={{width: '100%'}}>
+            <Header/>
             <h1>Would You Rather</h1>
             <div style={containerStyle}>
-                <div style={cardStyle}>
-                    <h2>Build our new app with JS</h2>
-                    <Button variant="contained">This One!</Button>
-                </div>
-                <div style={cardStyle}>
-                    <h2>Build with TS</h2>
-                    <Button variant="contained">No No, This One!</Button>
-                </div>
+                <PollOption pollData={{question: 'Build our new app with JS', id: 1}}/>
+                <PollOption pollData={{question: 'Build with TS', id:1}}/>
             </div>
         </div>
     )
